+++ Modulos/PuntosDeInteres/Dashboard.html
@@ -64,15 +64,9 @@
                                 <a class="link-dark d-inline-flex text-decoration-none rounded" href="./PuntosInteresMolde.html">Dashboard</a>
                             </li>     
                         <li class="mb-1">
-                            <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#home-collapse" aria-expanded="false">
-                            Puntos de Interes
-                            </button>
-                            <div class="collapse" id="home-collapse">
-                            <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
-                                <li><a href="./Gestion/PuntosInteresAlta.html" class="ms-3 link-dark d-inline-flex text-decoration-none rounded">Agregar Nuevo</a></li>
-                                <li><a href="./Gestion/PuntosInteresShow.html" class="ms-3 link-dark d-inline-flex text-decoration-none rounded">Mostrar Todos</a></li>
-                            </ul>
-                            </div>
+                            <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0">
+                            <a href="./PuntosInteresShow.html" class="link-dark text-decoration-none">Puntos de Interes</a>
+                            </button>  
                         </li>
                         <li class="mb-1">
                             <button class="btn btn-toggle  d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#dashboard-collapse" aria-expanded="false">
@@ -99,6 +93,7 @@
         
     </div>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <script src="../../formularios.js"></script>
     <script src="../../app.js"></script>
     <script src="../../bootstrap/js/bootstrap.bundle.js"></script>
 </body>
