+++ Modulos/PuntosDeInteres/Formularios/ServiciosEscencialesAlta.html
@@ -1,0 +1,24 @@
+<div class="row justify-content-center">
+    <div class="col-sm-6 mt-3">
+        <form action="" method="POST">
+            <h3 class="text-center">SERVICIOS ESENCIALES</h3>
+            <div class="mb-3">
+                <select id="TipoDeServicioEsencial" class="form-select" aria-label="Default select example">
+                    <option selected>Seleccionar Tipo</option>
+                    <option value="Hospitales">Hospitales</option>
+                    <option value="Farmacias">Farmacias</option>
+                    <option value="Cerrajerias">Cerrajerias</option>
+                    <option value="Estaciones de Servicio">Estaciones de Servicio</option>
+                    <option value="Seccionales">Seccionales</option>
+                </select>
+            </div>
+            
+        </form>
+    </div>
+</div>
+<div style="height: 60vh;">
+
+</div>
+<div class="row mb-3 justify-content-center">
+    <input type="button" class="w-50 btn btn-primary form-control" onclick="AltaDeServicioEscencial();" value="Registrar">
+</div>
