--- Modulos/PuntosDeInteres/PuntosInteresMolde.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../../bootstrap/css/bootstrap.css">
-    <title>Document</title>
-</head>
-<body>
-    <!--Contenedor -->
-    <div  class="container-fluid">
-        <div class="row">
-            <header class="rounded" style="border:2px solid black">
-                <div class="col-sm m-2">
-                    <form class="row justify-content-center" method="POST">
-                        <div class="col-3">
-                        <input type="password" class="form-control" id="inputPassword2" placeholder="Buscar">
-                        </div>
-                        <div class="col-auto">
-                        <input type="button" class="btn btn-primary" value="Buscar">
-                        </div>
-                    </form>
-                </div>
-            </header>    
-        </div>
-        
-            
-        
-        <!--Aside y Main -->
-        <div style="height: 90vh;" class="row">
-            <aside class="col-3 rounded" style="border:2px solid black;">
-                    <div class="flex-shrink-0 p-3 ms-3 bg-white" style="width: 280px;">
-                        <a href="/" class="d-flex align-items-center pb-3 mb-3 link-dark text-decoration-none border-bottom">
-                        <svg class="bi pe-none me-2" width="30" height="24"><use xlink:href="#bootstrap"></use></svg>
-                        <span class="fs-5 fw-semibold">Menu Gestor Turistico</span>
-                        </a>
-                        <ul class="list-unstyled ps-0">
-                            <li class="mb-1">
-                                <a class="link-dark d-inline-flex text-decoration-none rounded" href="./PuntosInteresMolde.html">Dashboard</a>
-                            </li>     
-                        <li class="mb-1">
-                            <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#home-collapse" aria-expanded="false">
-                            Puntos de Interes
-                            </button>
-                            <div class="collapse" id="home-collapse">
-                            <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
-                                <li><a href="./Gestion/PuntosInteresAlta.html" class="ms-3 link-dark d-inline-flex text-decoration-none rounded" data-bs-toggle="collapse" data-bs-target="#categorias-collapse">Agregar Nuevo</a></li>
-                                <div class="collapse" id="categorias-collapse">
-                                    <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
-                                        <li><a href="./Gestion/PuntosInteresAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Punto Interes</a></li>
-                                        <li><a href="./Gestion/Categorias/AlojamientoAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Alojamientos</a></li>
-                                        <li><a href="./Gestion/Categorias/GastronomiaAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Gastronomia</a></li>
-                                        <li><a href="./Gestion/Categorias/EspectaculosAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Espectaculos</a></li>
-                                        <li><a href="./Gestion/Categorias/ActividadesAlAireLibreAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Actividades al Aire Libre</a></li>
-                                        <li><a href="./Gestion/Categorias/ActividadesNocturnasAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Actividades Nocturnas</a></li>
-                                        <li><a href="./Gestion/Categorias/TransporteAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Transporte</a></li>
-                                        <li><a href="./Gestion/Categorias/ActividadesInfantilesAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Actividades Infantiles</a></li>
-                                        <li><a href="./Gestion/Categorias/ServiciosEscencialesAlta.html" class="ms-5 link-dark d-inline-flex text-decoration-none rounded">Servicios Escenciales</a></li>
-                                    </ul>
-                                    </div>
-                                <li><a href="./Gestion/PuntosInteresShow.html" class="ms-3 link-dark d-inline-flex text-decoration-none rounded">Mostrar Todos</a></li>
-                            </ul>
-                            </div>
-                        </li>
-                        <li class="mb-1">
-                            <button class="btn btn-toggle  d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#dashboard-collapse" aria-expanded="false">
-                            Gestion de Tours
-                            </button>
-                            <div class="collapse" id="dashboard-collapse">
-                            <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
-                                <li><a href="#" class="link-dark d-inline-flex text-decoration-none rounded">Agregar Nuevo</a></li>
-                                <li><a href="#" class="link-dark d-inline-flex text-decoration-none rounded">Mostrar Todos</a></li>
-                            </ul>
-                            </div>
-                        </li>
-                        
-                        <li class="border-top my-3"></li>
-                        
-                        </ul>
-                    </div>
-            </aside>
-            <main class="col-9 rounded" style="border:2px solid black;">
-
-            </main>
-        </div>
-            
-        
-    </div>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="../../app.js"></script>
-    <script src="../../bootstrap/js/bootstrap.bundle.js"></script>
-</body>
-</html>
