+++ index.html
@@ -29,7 +29,7 @@
                     <input type="text" class="form-control" placeholder="Contraseña">
                 </div>
                 <div class="mb-3">
-                    <a href="./Modulos/PuntosDeInteres/PuntosInteresMolde2.html">Avanzar --></a>
+                    <a href="./Modulos/PuntosDeInteres/Dashboard.html">Avanzar --></a>
                     <input type="text" class="btn btn-primary form-control" value="Ingresar">
                 </div>
             </form>
@@ -38,5 +38,6 @@
     </div>
     <script src="./bootstrap/js/bootstrap.bundle.js"></script>
     <script src="./app.js"></script>
+    <link rel="stylesheet" href="./Modulos/PuntosDeInteres/Formularios/ServiciosEscencialesAlta.html">
 </body>
 </html>
